# PERF: Avoid np.divmod in maybe_sequence_to_range